Testing parallelism with `Threads.@threads`
Eliminating the dynamics defects is one of the main chores in the optimisation problem. Each defect is formulated as an equality constraint (`== 0`).

I wanted to try to make the solver converge faster by evaluating the constraint function and Jacobian in parallel across multiple threads, but for some reason it makes it slower... Anyway, I am still committing this to a separate branch for posterity.

The number of threads to be used must be set before starting Julia. If not specified, Julia defaults to a single thread. In order to set e.g. 4 threads run:
```export JULIA_NUM_THREADS=4```
(See https://docs.julialang.org/en/v1/manual/parallel-computing/#man-multithreading-1.)

I suspect the reason why it becomes slower may be due to the body of the loops not being demanding enough computationally-wise to justify the extra time taken by the _context switching_ between threads... I am not an expert in parallel computing though, so I am not entirely sure about this one.